Add loading of YUV textures
Changes existing motion vector computation in
computeSumOfAbsoluteDifference to use the "R" channel since the YUV
textures' Y channel is loaded into a R8Unorm texture.

Note that the current implementation was using the G channel of an
RGBA8 texture, pretending it were Y.

Also changes IOSurfaces to be allocated with storage mode Managed
as this is required by the Metal validation layers.